lib/igt_pm: Convert suspended_time from int to uint64_t
Suspended time is in milliseconds printed from u64 variable in kernel.
Convert to a more appropriate variable type.

v2: - Use PRInt macros instead of %ld (Kamil)
    - Explicit time units (Kamil)
    - s/suspend/suspended in the debug msg
v3: - s/miliseconds/milliseconds (Francois)

Cc: Francois Dugast <francois.dugast@intel.com>
Cc: Swati Sharma <swati2.sharma@intel.com>
Suggested-by: Kamil Konieczny <kamil.konieczny@linux.intel.com>

+++ lib/igt_pm.h
@@ -93,7 +93,7 @@ void igt_pm_restore_pci_card_runtime_pm(void);
 void igt_pm_print_pci_card_runtime_status(void);
 bool i915_is_slpc_enabled_gt(int drm_fd, int gt);
 bool i915_is_slpc_enabled(int drm_fd);
-int igt_pm_get_runtime_suspended_time(struct pci_device *pci_dev);
+uint64_t igt_pm_get_runtime_suspended_time(struct pci_device *pci_dev);
 uint64_t igt_pm_get_runtime_active_time(struct pci_device *pci_dev);
 int igt_pm_get_runtime_usage(struct pci_device *pci_dev);
 void igt_pm_ignore_slpc_efficient_freq(int i915, int gtfd, bool val);
